+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/SpannerPool.java
@@ -165,6 +165,7 @@ public class SpannerPool {
     private final String clientCertificateKey;
     private final boolean isExperimentalHost;
     private final Boolean enableDirectAccess;
+    private final String universeDomain;
 
     @VisibleForTesting
     static SpannerPoolKey of(ConnectionOptions options) {
@@ -200,6 +201,7 @@ public class SpannerPool {
       this.clientCertificateKey = options.getClientCertificateKey();
       this.isExperimentalHost = options.isExperimentalHost();
       this.enableDirectAccess = options.isEnableDirectAccess();
+      this.universeDomain = options.getUniverseDomain();
     }
 
     @Override
@@ -226,7 +228,8 @@ public class SpannerPool {
           && Objects.equals(this.clientCertificate, other.clientCertificate)
           && Objects.equals(this.clientCertificateKey, other.clientCertificateKey)
           && Objects.equals(this.isExperimentalHost, other.isExperimentalHost)
-          && Objects.equals(this.enableDirectAccess, other.enableDirectAccess);
+          && Objects.equals(this.enableDirectAccess, other.enableDirectAccess)
+          && Objects.equals(this.universeDomain, other.universeDomain);
     }
 
     @Override
@@ -249,7 +252,8 @@ public class SpannerPool {
           this.clientCertificate,
           this.clientCertificateKey,
           this.isExperimentalHost,
-          this.enableDirectAccess);
+          this.enableDirectAccess,
+          this.universeDomain);
     }
   }
 
@@ -418,6 +422,9 @@ public class SpannerPool {
     }
     if (key.enableDirectAccess != null) {
       builder.setEnableDirectAccess(key.enableDirectAccess);
+    }
+    if (key.universeDomain != null) {
+      builder.setUniverseDomain(key.universeDomain);
     }
     if (options.getConfigurator() != null) {
       options.getConfigurator().configure(builder);